+++ SocialMediaApp/SocialMediaApp/src/main/java/com/Licenta/SocialMediaApp/Repository/ContentRepository.java
@@ -3,5 +3,5 @@ package com.Licenta.SocialMediaApp.Repository;
 import com.Licenta.SocialMediaApp.Model.Content;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContentRepository extends JpaRepository<Content, Integer> {
+public interface ContentRepository extends JpaRepository<Content, Long> {
 }
