+++ SocialMediaApp/SocialMediaApp/src/main/java/com/Licenta/SocialMediaApp/Service/FriendsListService.java
@@ -6,9 +6,9 @@ import com.Licenta.SocialMediaApp.Model.User;
 import java.util.List;
 
 public interface FriendsListService {
-    int countNrOfFriends(int userId);
-    boolean isFriendshipExists(int userId1, int userId2);
-    List<User> findFriendsByUserId(int userId);
+    int countNrOfFriends(Long userId);
+    boolean isFriendshipExists(Long userId1, Long userId2);
+    List<User> findFriendsByUserId(Long userId);
     FriendsList createFriendsList(FriendsList friendsList);
-    void deleteFriend(String jwt, int userId);
+    void deleteFriend(String jwt, Long userId);
 }
