+++ SocialMediaApp/SocialMediaApp/src/main/java/com/Licenta/SocialMediaApp/Service/LikeService.java
@@ -3,7 +3,7 @@ package com.Licenta.SocialMediaApp.Service;
 import com.Licenta.SocialMediaApp.Model.Like;
 
 public interface LikeService {
-    Like addLike(String jwt, int postId);
-    void deleteLike(String jwt, int postId);
-    boolean checkUserLikedPost(String jwt, int postId);
+    Like addLike(String jwt, Long postId);
+    void deleteLike(String jwt, Long postId);
+    boolean checkUserLikedPost(String jwt, Long postId);
 }
